VirtualSlotCar: Changed spline coefficient computation to double precision, allowed setting of integration step size for curvature computations.
git-svn-id: https://svn.code.sf.net/p/jaer/code/trunk/host/java@2128 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/ch/unizh/ini/jaer/projects/virtualslotcar/SlotcarFrame.java
@@ -310,7 +310,7 @@ public class SlotcarFrame extends javax.swing.JFrame {
     }//GEN-LAST:event_DeletePointsButtonActionPerformed
 
     private void stepsizeValueActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_stepsizeValueActionPerformed
-        double step = Double.parseDouble(stepsizeValue.getText());
+        float step = Float.parseFloat(stepsizeValue.getText());
         if ((step > 0) && (step <1))
             EditorPanel.setStepSize(step);
         if (step >= 1)
